Need some comment for first deploy
I spent a few days figuring out why following all the steps in this article didn't work. The reason was the healthcheck for litestream always fails if no db exists and no replica exists, which is true for the first deployment.
My workaround, which I don't think is something I'd advise to do, was to change the healthcheck test to 
  healthcheck:
      test: /usr/local/bin/litestream version
which will pass first use.